refactor(model): 도메인 패키지로 리팩토링
- 모델 패키지 안에 도메인 + 서비스 패키지를 같이 놔두기
