+++ src/main/java/lotto/model/domain/LottoPrize.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 import java.util.HashMap;
 import java.util.Map;
