+++ src/main/java/lotto/model/domain/UserStatus.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 public enum UserStatus {
     DEFAULT,
